+++ apps/web/src/app/[locale]/settings/profile/page.tsx
@@ -20,11 +20,7 @@ const profileFormSchema = z.object({
     .max(30, {
       message: "Username must not be longer than 30 characters.",
     }),
-  email: z
-    .string({
-      required_error: "Please select an email to display.",
-    })
-    .email(),
+  email: z.string().min(1, { message: "Please select an email to display." }).email(),
   bio: z.string().max(160).min(4),
 });